Omero Sudo (omrusrsudo) bug
When an Omero user (without administrator previledges) was running the macro, the Ext.sudo(omrusrsudo) command was always running, even if the user had not put a username at the respective field. 

Now this is corrected. If the "run as a different user" field is empty, then the omero orperations are run as the logged in user.